pyproject.toml
[project]
name = "aria-blue"
version = "3.0.0"
description = "Autonomous AI agent platform with multi-model LLM routing, modular skills, and multi-agent orchestration"
readme = "README.md"
requires-python = ">=3.13"
license = {file = "LICENSE"}
authors = [
    {name = "Najia-afk"}
]
keywords = ["ai", "agent", "autonomous", "llm", "moltbook", "multi-agent", "aria"]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "httpx>=0.25.0",
    "asyncpg>=0.29.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.0",
    "litellm>=1.55.0",
    "apscheduler>=4.0.0a1",
    "aiohttp>=3.9",
    "sqlalchemy[asyncio]>=2.0.30",
    "alembic>=1.13.0",
    "prometheus-client>=0.21.0",
    "websockets>=13.0",
    "uvicorn[standard]>=0.30.0",
    "fastapi>=0.115.0",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "requests>=2.31.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "freezegun>=1.4.0",
    "locust>=2.29.0",
    "memray>=1.14.0",
    "py-spy>=0.3.14",
    "pyupgrade>=3.17.0",
    "testcontainers>=4.0.0",
]

docker = [
    "docker>=7.0.0",
]

all = [
    "aria-blue[dev]",
    "aria-blue[docker]",
]

[project.urls]
Homepage = "https://datascience-adventure.xyz"
Repository = "https://github.com/Najia-afk/Aria_moltbot"
Issues = "https://github.com/Najia-afk/Aria_moltbot/issues"
License = "https://datascience-adventure.xyz/contact"

[project.scripts]
aria = "aria_mind.cli:main"
aria-health = "aria_mind.cli:health_check"
aria-engine = "aria_engine.entrypoint:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["aria_mind", "aria_skills", "aria_agents", "aria_engine"]

# ============================================================================
# pytest
# ============================================================================

[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "unit: fast unit tests",
    "integration: requires mocked services",
    "docker: only runs in Docker container",
    "load: load/performance tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

# ============================================================================
# coverage
# ============================================================================

[tool.coverage.run]
source = ["src", "aria_skills", "aria_mind", "aria_models", "aria_agents", "aria_engine"]
branch = true
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
show_missing = true

# ============================================================================
# black
# ============================================================================

[tool.black]
line-length = 100
target-version = ["py313"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hatch
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

# ============================================================================
# ruff
# ============================================================================

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex
    "ARG001", # unused function argument
]

[tool.ruff.lint.isort]
known-first-party = ["aria_mind", "aria_skills", "aria_agents", "aria_engine"]

# ============================================================================
# mypy
# ============================================================================

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
exclude = [
    "tests/",
    "build/",
    "dist/",
]

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true
